feat: invite classes based on type
