+++ packages/discord.js/src/client/websocket/handlers/INVITE_CREATE.js
@@ -14,7 +14,7 @@ module.exports = (client, { d: data }) => {
    * Emitted when an invite is created.
    * <info>This event requires the {@link PermissionFlagsBits.ManageChannels} permission for the channel.</info>
    * @event Client#inviteCreate
-   * @param {Invite} invite The invite that was created
+   * @param {GuildInvite} invite The invite that was created
    */
   client.emit(Events.InviteCreate, invite);
 };
