+++ packages/discord.js/src/index.js
@@ -114,6 +114,7 @@ exports.BaseGuildEmoji = require('./structures/BaseGuildEmoji.js').BaseGuildEmoj
 exports.BaseGuildTextChannel = require('./structures/BaseGuildTextChannel.js').BaseGuildTextChannel;
 exports.BaseGuildVoiceChannel = require('./structures/BaseGuildVoiceChannel.js').BaseGuildVoiceChannel;
 exports.BaseInteraction = require('./structures/BaseInteraction.js').BaseInteraction;
+exports.BaseInvite = require('./structures/BaseInvite.js').BaseInvite;
 exports.BaseSelectMenuComponent = require('./structures/BaseSelectMenuComponent.js').BaseSelectMenuComponent;
 exports.ButtonComponent = require('./structures/ButtonComponent.js').ButtonComponent;
 exports.ButtonInteraction = require('./structures/ButtonInteraction.js').ButtonInteraction;
@@ -138,12 +139,14 @@ exports.Embed = require('./structures/Embed.js').Embed;
 exports.Emoji = require('./structures/Emoji.js').Emoji;
 exports.Entitlement = require('./structures/Entitlement.js').Entitlement;
 exports.ForumChannel = require('./structures/ForumChannel.js').ForumChannel;
+exports.GroupDMInvite = require('./structures/GroupDMInvite.js').GroupDMInvite;
 exports.Guild = require('./structures/Guild.js').Guild;
 exports.GuildAuditLogs = require('./structures/GuildAuditLogs.js').GuildAuditLogs;
 exports.GuildAuditLogsEntry = require('./structures/GuildAuditLogsEntry.js').GuildAuditLogsEntry;
 exports.GuildBan = require('./structures/GuildBan.js').GuildBan;
 exports.GuildChannel = require('./structures/GuildChannel.js').GuildChannel;
 exports.GuildEmoji = require('./structures/GuildEmoji.js').GuildEmoji;
+exports.GuildInvite = require('./structures/GuildInvite.js').GuildInvite;
 exports.GuildMember = require('./structures/GuildMember.js').GuildMember;
 exports.GuildOnboarding = require('./structures/GuildOnboarding.js').GuildOnboarding;
 exports.GuildOnboardingPrompt = require('./structures/GuildOnboardingPrompt.js').GuildOnboardingPrompt;
@@ -162,7 +165,6 @@ exports.InteractionCallbackResponse =
   require('./structures/InteractionCallbackResponse.js').InteractionCallbackResponse;
 exports.InteractionCollector = require('./structures/InteractionCollector.js').InteractionCollector;
 exports.InteractionWebhook = require('./structures/InteractionWebhook.js').InteractionWebhook;
-exports.Invite = require('./structures/Invite.js').Invite;
 exports.InviteGuild = require('./structures/InviteGuild.js').InviteGuild;
 exports.MediaChannel = require('./structures/MediaChannel.js').MediaChannel;
 exports.MentionableSelectMenuComponent =
