+++ packages/discord.js/src/managers/GuildInviteManager.js
@@ -4,7 +4,7 @@ const { Collection } = require('@discordjs/collection');
 const { Routes } = require('discord-api-types/v10');
 const { CachedManager } = require('./CachedManager.js');
 const { DiscordjsError, ErrorCodes } = require('../errors/index.js');
-const { Invite } = require('../structures/Invite.js');
+const { GuildInvite } = require('../structures/GuildInvite.js');
 const { resolveInviteCode } = require('../util/DataResolver.js');
 
 /**
@@ -13,7 +13,7 @@ const { resolveInviteCode } = require('../util/DataResolver.js');
  */
 class GuildInviteManager extends CachedManager {
   constructor(guild, iterable) {
-    super(guild.client, Invite, iterable);
+    super(guild.client, GuildInvite, iterable);
 
     /**
      * The guild this Manager belongs to
@@ -24,7 +24,7 @@ class GuildInviteManager extends CachedManager {
 
   /**
    * The cache of this Manager
-   * @type {Collection<string, Invite>}
+   * @type {Collection<string, GuildInvite>}
    * @name GuildInviteManager#cache
    */
 
@@ -33,7 +33,7 @@ class GuildInviteManager extends CachedManager {
   }
 
   /**
-   * Data that resolves to give an Invite object. This can be:
+   * Data that resolves to give a `GuildInvite`. This can be:
    * * An invite code
    * * An invite URL
    * @typedef {string} InviteResolvable
@@ -53,12 +53,12 @@ class GuildInviteManager extends CachedManager {
    */
 
   /**
-   * Resolves an InviteResolvable to an Invite object.
+   * Resolves an InviteResolvable to a `GuildInvite` object.
    * @method resolve
    * @memberof GuildInviteManager
    * @instance
    * @param {InviteResolvable} invite The invite resolvable to resolve
-   * @returns {?Invite}
+   * @returns {?GuildInvite}
    */
 
   /**
@@ -89,7 +89,7 @@ class GuildInviteManager extends CachedManager {
   /**
    * Fetches invite(s) from Discord.
    * @param {InviteResolvable|FetchInviteOptions|FetchInvitesOptions} [options] Options for fetching guild invite(s)
-   * @returns {Promise<Invite|Collection<string, Invite>>}
+   * @returns {Promise<GuildInvite|Collection<string, GuildInvite>>}
    * @example
    * // Fetch all invites from a guild
    * guild.invites.fetch()
@@ -170,7 +170,7 @@ class GuildInviteManager extends CachedManager {
    * Create an invite to the guild from the provided channel.
    * @param {GuildInvitableChannelResolvable} channel The options for creating the invite from a channel.
    * @param {InviteCreateOptions} [options={}] The options for creating the invite from a channel.
-   * @returns {Promise<Invite>}
+   * @returns {Promise<GuildInvite>}
    * @example
    * // Create an invite to a selected channel
    * guild.invites.create('599942732013764608')
@@ -196,7 +196,7 @@ class GuildInviteManager extends CachedManager {
       },
       reason,
     });
-    return new Invite(this.client, invite);
+    return new GuildInvite(this.client, invite);
   }
 
   /**
