+++ packages/discord.js/typings/index.test-d.ts
@@ -214,6 +214,8 @@ import {
   PrimaryButtonBuilder,
   resolveColor,
   createComponentBuilder,
+  GuildInvite,
+  Invite,
 } from './index.js';
 import { expectAssignable, expectNotAssignable, expectNotType, expectType } from 'tsd';
 import type { ContextMenuCommandBuilder, ChatInputCommandBuilder } from '@discordjs/builders';
@@ -248,6 +250,10 @@ if (client.isReady()) {
 } else {
   expectType<Client>(client);
 }
+
+expectType<Promise<Invite>>(client.fetchInvite('https://discord.gg/djs'));
+expectType<Promise<Invite<true>>>(client.fetchInvite('https://discord.gg/djs', { withCounts: true }));
+expectNotType<Promise<Invite<true>>>(client.fetchInvite('https://discord.gg/djs', { withCounts: false }));
 
 const testGuildId = '222078108977594368'; // DJS
 const testUserId = '987654321098765432'; // example id
@@ -391,8 +397,15 @@ client.on('interactionCreate', async interaction => {
   }
 });
 
-client.on('inviteCreate', ({ client }) => expectType<Client<true>>(client));
-client.on('inviteDelete', ({ client }) => expectType<Client<true>>(client));
+client.on('inviteCreate', invite => {
+  expectType<GuildInvite>(invite);
+  expectType<Client<true>>(invite.client);
+});
+
+client.on('inviteDelete', invite => {
+  expectType<GuildInvite>(invite);
+  expectType<Client<true>>(invite.client);
+});
 
 // This is to check that stuff is the right type
 declare const assertIsMessage: (m: Promise<Message>) => void;
